+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Features:
 Fixes:
 
 * Don't reload edit-forms if some entry-status changed [#99](https://github.com/O-X-L/ansible-webui/issues/99)
+* Better handling for job-execution log-redirect [#97](https://github.com/O-X-L/ansible-webui/issues/97)
 
 Project: